R markdown script and html
R is used as a tool to explore the dataset. Some data preprocessing is
done: verification of datatypes of all columns, remove outliers and
duplicates. simple linear regression is used to learn the linearity
between the price and the surface.
However, different patterns are observed according to different cities,
some plots are made to dig into these subsets.